Started to add remove document function

+++ DocumentsSecurity/DocumentsSecurity/AddProjectDialog.cs
@@ -37,7 +37,7 @@ namespace DocumentsSecurity
         {
             bool isAllOk = true;
             //id isn't important, it's generating at runtime
-            long id = 0;
+            int id = 0;
 
             string customer = DocumentProjectCustomerTextBox.Text;
             customer = customer == null ? "" : customer;
